bug fixes add infect

+++ TPTBox/segmentation/TotalVibeSeg/auto_download.py
@@ -75,7 +75,7 @@ def _download_weights(idx=85, addendum="", first=True) -> None:
         addendum_download(idx)
 
 
-def _download(weights_url, weights_dir, text="") -> None:
+def _download(weights_url, weights_dir, text="", is_zip=True) -> None:
     try:
         # Retrieve file size
         with urllib.request.urlopen(str(weights_url)) as response:
@@ -95,11 +95,11 @@ def _download(weights_url, weights_dir, text="") -> None:
         zip_path = weights_dir.parent / Path(weights_url).name
         # Download the file
         urllib.request.urlretrieve(str(weights_url), zip_path, reporthook=update_progress)
-
-    print(f"Extracting {text}...")
-    with zipfile.ZipFile(zip_path, "r") as zip_ref:
-        zip_ref.extractall(weights_dir)
-    os.remove(zip_path)  # noqa: PTH107
+    if is_zip:
+        print(f"Extracting {text}...")
+        with zipfile.ZipFile(zip_path, "r") as zip_ref:
+            zip_ref.extractall(weights_dir)
+        os.remove(zip_path)  # noqa: PTH107
 
 
 def addendum_download(idx):
